+++ src/components/shared/RentInformationTable/AddRentPaymentModal/AddRentPaymentModal.tsx
@@ -4,10 +4,12 @@ import RentInformationUtils from '../../../../utils/rent/rent.utils';
 import { ValidationConstraint } from '../../../../utils/validation/constraints';
 import Validator from '../../../../utils/validation/validator';
 import CurrencyInputField from '_/components/form/CurrencyInputField/CurrencyInputField';
+import FileSelect from '_/components/form/FileSelect/FileSelect';
 import StandardDateField from '_/components/form/StandardDateField/StandardDateField';
 import GenericModal from '_/components/generic/GenericModal/GenericModal';
 import useFormValidation from '_/hooks/useFormValidation/useFormValidation';
 import useResident from '_/hooks/useResident/useResident';
+import { DocumentType } from '_/models/document';
 import { RentInformation } from '_/models/resident/rent';
 import { Resident } from '_/models/resident/resident';
 import { CurrencyInCents } from '_/utils/currency/currency.utils';
@@ -47,13 +49,18 @@ interface AddRentPaymentInput {
    * Date of the payment
    */
   paymentDate: Date;
+
+  /**
+   * Path to the bank transfer document
+   */
+  bankTransferFile: string;
 }
 
 /**
  * Modal that provides functionality to insert payment information
  */
 function AddRentPaymentModal(props: AddRentPaymentModalProps): JSX.Element {
-  const { addRentPayment } = useResident(props.resident.id);
+  const { addRentPayment, addDocument } = useResident(props.resident.id);
 
   const {
     formInput,
@@ -65,17 +72,35 @@ function AddRentPaymentModal(props: AddRentPaymentModalProps): JSX.Element {
     formValidator: new Validator<AddRentPaymentInput>({
       paymentAmount: ValidationConstraint.Currency,
       paymentDate: ValidationConstraint.Defined,
+      bankTransferFile: ValidationConstraint.DefinedFile,
     }),
     defaultFormInput: {
       paymentAmount: RentInformationUtils.getAmountToPay(props.rentInformation),
       paymentDate: new Date().toUTC(),
+      bankTransferFile: undefined,
     },
     onSubmitSuccess: (values) => {
-      addRentPayment({
-        dueDate: props.rentInformation.dueDate,
-        ...values,
-      });
-      props.onCloseModal();
+      void window.ipcAPI.persistence
+        .uploadDocument(values.bankTransferFile, {
+          type: 'resident',
+          residentId: props.resident.id,
+        })
+        .then((bankTransferDocumentId) => {
+          addDocument({
+            name: `Mietzahlung ${props.rentInformation.dueDate.toString()}`,
+            type: DocumentType.BankTransfer,
+            creationDate: values.paymentDate,
+            subjectDate: props.rentInformation.dueDate,
+            id: bankTransferDocumentId,
+          });
+          addRentPayment({
+            dueDate: props.rentInformation.dueDate,
+            paymentAmount: values.paymentAmount,
+            paymentDate: values.paymentDate,
+            bankTransferDocumentId,
+          });
+          props.onCloseModal();
+        });
     },
     submitButtonLabel: 'Hinzufügen',
   });
@@ -109,6 +134,15 @@ function AddRentPaymentModal(props: AddRentPaymentModalProps): JSX.Element {
             value={formInput.paymentDate}
             onChange={formInputSetters.paymentDate}
             errorMessage={formErrors.paymentDate}
+          />
+        </Grid>
+        <Grid item xs={12}>
+          <FileSelect
+            id="bankTransferFile"
+            label="Überweisung"
+            value={formInput.bankTransferFile}
+            onChange={formInputSetters.bankTransferFile}
+            errorMessage={formErrors.bankTransferFile}
           />
         </Grid>
       </Grid>