+++ src/ipc/modules/persistence/uploadDocument/uploadDocument.spec.ts
@@ -1,20 +1,25 @@
 import fs from 'fs';
 import path from 'path';
+import { v4 } from 'uuid';
 import { DocumentTarget } from '../../../utils/persistence/documentTarget';
 import * as getAppDataDirectoryModule from '../../../utils/persistence/getAppDataDirectory';
 import uploadDocument from './uploadDocument';
 
+jest.mock('uuid', () => ({
+  v4: jest.fn(),
+}));
+
 describe('uploadDocument', () => {
   const appDataDir = 'appData';
   const uploadedFile = 'test/file.txt';
-  const fileName = 'test.txt';
   const target: DocumentTarget = {
     type: 'resident',
     residentId: 'resident1',
   };
+  const expectedDocumentId = 'uploaded-document-id';
 
-  const destinationDir = path.join(appDataDir, 'documents/residents/resident1');
-  const destinationFile = path.join(destinationDir, fileName);
+  const destinationDir = path.join(appDataDir, `documents/residents/${target.residentId}`);
+  const destinationFile = path.join(destinationDir, `${expectedDocumentId}.txt`);
 
   let existsSyncMock: jest.SpyInstance;
   let mkdirSyncMock: jest.SpyInstance;
@@ -28,6 +33,8 @@ describe('uploadDocument', () => {
     existsSyncMock = jest.spyOn(fs, 'existsSync').mockReturnValue(true);
     mkdirSyncMock = jest.spyOn(fs, 'mkdirSync').mockImplementation();
     copyFileSyncMock = jest.spyOn(fs, 'copyFileSync').mockImplementation();
+
+    (v4 as jest.Mock).mockReturnValue(expectedDocumentId);
   });
 
   afterEach(() => {
@@ -36,7 +43,7 @@ describe('uploadDocument', () => {
 
   it('should copy file correctly', () => {
     // Act
-    uploadDocument(uploadedFile, fileName, target);
+    const documentId = uploadDocument(uploadedFile, target);
 
     // Assert
     expect(copyFileSyncMock).toHaveBeenCalledTimes(1);
@@ -44,6 +51,7 @@ describe('uploadDocument', () => {
       uploadedFile,
       destinationFile,
     );
+    expect(documentId).toBe(expectedDocumentId);
   });
 
   it('should create dir if not existing', () => {
@@ -51,7 +59,7 @@ describe('uploadDocument', () => {
     existsSyncMock.mockReturnValue(false);
 
     // Act
-    uploadDocument(uploadedFile, fileName, target);
+    uploadDocument(uploadedFile, target);
 
     // Assert
     expect(mkdirSyncMock).toHaveBeenCalledTimes(1);
@@ -65,7 +73,7 @@ describe('uploadDocument', () => {
     existsSyncMock.mockReturnValue(true);
 
     // Act
-    uploadDocument(uploadedFile, fileName, target);
+    uploadDocument(uploadedFile, target);
 
     // Assert
     expect(mkdirSyncMock).toHaveBeenCalledTimes(0);