+++ src/views/ResidentView/ResidentInformation/DocumentInformation/UploadDocumentModal/UploadDocumentModal.spec.tsx
@@ -10,7 +10,6 @@ import { setRecoil } from 'recoil-nexus';
 import UploadDocumentModal from './UploadDocumentModal';
 import useResident, * as useResidentModule from '_/hooks/useResident/useResident';
 import { LinkedDocument, DocumentType } from '_/models/document';
-
 import residentState from '_/states/resident/resident.state';
 import ReactTestWrapper from '_/test/ReactTestWrapper';
 import ResidentBuilder from '_/test/builders/resident.builder';
@@ -23,6 +22,7 @@ describe('UploadDocumentModal', () => {
   const onCloseModalMock = jest.fn();
 
   const resident = new ResidentBuilder().build();
+  const documentId = 'uploaded-document-id';
 
   const validInputValues = {
     name: 'Test Document',
@@ -78,6 +78,8 @@ describe('UploadDocumentModal', () => {
   beforeEach(() => {
     jest.spyOn(useResidentModule, 'default').mockReturnValue(useResidentMock);
 
+    mockedIpcAPIFunctions.persistence.uploadDocument.mockResolvedValue(documentId);
+
     baseElement = render(
       <ReactTestWrapper
         initializationFunction={() => {
@@ -117,15 +119,13 @@ describe('UploadDocumentModal', () => {
 
   test('should update resident on submit', async () => {
     // Arrange
-    const expectedDocument: Omit<LinkedDocument, 'id'> = {
+    const expectedDocument: LinkedDocument = {
       name: validInputValues.name,
       creationDate: validInputValues.date,
       subjectDate: validInputValues.date,
       type: DocumentType.CoverLetter,
+      id: documentId,
     };
-    mockedIpcAPIFunctions.persistence.uploadDocument.mockResolvedValue(
-      undefined,
-    );
 
     await inputToForm(validInputValues);
 
@@ -142,7 +142,6 @@ describe('UploadDocumentModal', () => {
       mockedIpcAPIFunctions.persistence.uploadDocument,
     ).toHaveBeenLastCalledWith(
       validInputValues.file,
-      expect.stringMatching(/^.*\.txt$/),
       {
         type: 'resident',
         residentId: resident.id,