+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Rent payments
     - Added payment amount default value to the add rent payments form
+    - Added bank transfer document
 
 ## v0.2.0 - 2024-08-21